Move test/node wait for resources functions to a separate module
